% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukds_download.R
\name{ukds_download}
\alias{ukds_download}
\title{Download datasets from the UK Data Service}
\usage{
ukds_download(file_id, org = getOption("ukds_org"),
  user = getOption("ukds_user"), password = getOption("ukds_password"),
  reset = FALSE, download_dir = "ukds_data", msg = TRUE, convert = TRUE)
}
\arguments{
\item{file_id}{The unique identifier (or optionally a vector of these identifiers)
for the dataset(s) to be downloaded (see details).}

\item{org, email, password}{Your UK Data Service organization, username, and password (see details)}

\item{reset}{If TRUE, you will be asked to re-enter your organization, username, and password.}

\item{download_dir}{The directory (relative to your working directory) to
which files from the UK Data Service will be downloaded.}

\item{msg}{If TRUE, outputs a message showing which data set is being downloaded.}

\item{convert}{If TRUE, converts downloaded file(s) to .RData format.}
}
\value{
The function returns downloaded files.
}
\description{
\code{ukds_download} provides a programmatic and reproducible means to download datasets 
  from the UK Data Service's data archive
}
\details{
To avoid requiring others to edit your scripts to insert their own email and  
 password or to force them to do so interactively, the default is set to fetch 
 this information from the user's .Rprofile.  Before running \code{ukds_download}, 
 then, you should be sure to add these options to your .Rprofile substituting your 
 info for the example below:

   \code{
       options("ukds_org" = "UK Data Service",
               "ukds_user" = "ukf0000000000",
               "ukds_password" = "password123!")
   }
}
\examples{
\dontrun{
 ukds_download(file_id = c())
}

}
